Remove backport of functools.total_ordering
This backport was needed to support Python 2.6
