+++ lib/spack/spack/concretize.py
@@ -16,13 +16,12 @@ TODO: make this customizable and allow users to configure
 """
 from __future__ import print_function
 
+import functools
 import os.path
 import platform
 import tempfile
 from contextlib import contextmanager
 from itertools import chain
-
-from functools_backport import reverse_order
 
 import archspec.cpu
 
@@ -46,6 +45,23 @@ from spack.version import Version, VersionList, VersionRange, ver
 
 #: impements rudimentary logic for ABI compatibility
 _abi = llnl.util.lang.Singleton(lambda: spack.abi.ABI())
+
+
+@functools.total_ordering
+class reverse_order(object):
+    """Helper for creating key functions.
+
+       This is a wrapper that inverts the sense of the natural
+       comparisons on the object.
+    """
+    def __init__(self, value):
+        self.value = value
+
+    def __eq__(self, other):
+        return other.value == self.value
+
+    def __lt__(self, other):
+        return other.value < self.value
 
 
 class Concretizer(object):
@@ -130,11 +146,11 @@ class Concretizer(object):
 
         # Use a sort key to order the results
         return sorted(usable, key=lambda spec: (
-            not spec.external,                            # prefer externals
-            pref_key(spec),                               # respect prefs
-            spec.name,                                    # group by name
-            reverse_order(spec.versions),                 # latest version
-            spec                                          # natural order
+            not spec.external,             # prefer externals
+            pref_key(spec),                # respect prefs
+            spec.name,                     # group by name
+            reverse_order(spec.versions),  # latest version
+            spec                           # natural order
         ))
 
     def choose_virtual_or_external(self, spec):
